refactor(pages): remove TransactionsList component
